+++ .github/workflows/ci.yml
@@ -194,7 +194,7 @@ jobs:
       run: cargo $CMD test -p winit-common --features core-foundation --no-run
 
     - name: Test winit Common (XKB)
-      if: contains(matrix.platform.target, 'linux')
+      if: contains(matrix.platform.target, 'linux-gnu')
       run: cargo $CMD test -p winit-common --features xkb,x11,wayland --no-run
 
     - name: Test winit Orbital
@@ -209,6 +209,10 @@ jobs:
     - name: Test winit Win32
       if: contains(matrix.platform.target, 'windows')
       run: cargo $CMD test -p winit-win32 --target=${{ matrix.platform.target }}
+
+    - name: Test winit Wayland
+      if: contains(matrix.platform.target, 'linux-gnu')
+      run: cargo $CMD test -p winit-wayland --target=${{ matrix.platform.target }}
 
     # Test only on Linux x86_64, so we avoid spending unnecessary CI hours.
     - name: Test dpi crate
